+++ jpa-hibernate/src/main/java/com/example/jpa/JPASimple.java
@@ -34,7 +34,7 @@ public class JPASimple {
 		entityManager.persist(np);
 		transaction.commit();
 
-		int id = np.getId();
+		long id = np.getId();
 		System.out.println("Created Person with Id " + id);
 	}
